requirements.txt
tensorflow>=2.2.0
tensorflow-datasets>=3.1.0
numpy>=1.16.2
tf_siren>=0.0.4
Pillow
